+++ relay-server/src/actors/envelopes.rs
@@ -8,7 +8,7 @@ use chrono::Utc;
 use relay_base_schema::project::ProjectKey;
 use relay_config::{Config, HttpEncoding};
 use relay_event_schema::protocol::ClientReport;
-use relay_metrics::{Aggregator, Bucket, MergeBuckets};
+use relay_metrics::Bucket;
 use relay_quotas::Scoping;
 use relay_statsd::metric;
 use relay_system::{Addr, FromMessage, NoResponse};
@@ -28,6 +28,8 @@ use crate::extractors::{PartialDsn, RequestMeta};
 use crate::http::{HttpError, Request, RequestBuilder, Response};
 use crate::statsd::RelayHistograms;
 use crate::utils::ManagedEnvelope;
+
+use super::processor::EncodeMetrics;
 
 /// Error created while handling [`SendEnvelope`].
 #[derive(Debug, thiserror::Error)]
@@ -147,8 +149,6 @@ pub struct SendMetrics {
     pub buckets: Vec<Bucket>,
     /// Scoping information for the metrics.
     pub scoping: Scoping,
-    /// The key of the logical partition to send the metrics to.
-    pub partition_key: Option<u64>,
 }
 
 /// Dispatch service for generating and submitting Envelopes.
@@ -196,7 +196,6 @@ impl FromMessage<SendMetrics> for EnvelopeManager {
 #[derive(Debug)]
 pub struct EnvelopeManagerService {
     config: Arc<Config>,
-    aggregator: Addr<Aggregator>,
     enveloper_processor: Addr<EnvelopeProcessor>,
     project_cache: Addr<ProjectCache>,
     test_store: Addr<TestStore>,
@@ -209,7 +208,6 @@ impl EnvelopeManagerService {
     /// Creates a new instance of the [`EnvelopeManager`] service.
     pub fn new(
         config: Arc<Config>,
-        aggregator: Addr<Aggregator>,
         enveloper_processor: Addr<EnvelopeProcessor>,
         project_cache: Addr<ProjectCache>,
         test_store: Addr<TestStore>,
@@ -217,7 +215,6 @@ impl EnvelopeManagerService {
     ) -> Self {
         Self {
             config,
-            aggregator,
             enveloper_processor,
             project_cache,
             test_store,
@@ -238,7 +235,7 @@ impl EnvelopeManagerService {
         &self,
         mut envelope: Box<Envelope>,
         scoping: Scoping,
-        partition_key: Option<String>,
+        partition_key: Option<u64>,
     ) -> Result<(), SendEnvelopeError> {
         #[cfg(feature = "processing")]
         {
@@ -280,7 +277,7 @@ impl EnvelopeManagerService {
             http_encoding: self.config.http_encoding(),
             response_sender: tx,
             project_key: scoping.project_key,
-            partition_key,
+            partition_key: partition_key.map(|k| k.to_string()),
         };
 
         if let HttpEncoding::Identity = request.http_encoding {
@@ -300,9 +297,13 @@ impl EnvelopeManagerService {
         let SubmitEnvelope { mut envelope } = message;
 
         let scoping = envelope.scoping();
+        let partition_key = envelope.partition_key();
 
         let inner_envelope = envelope.take_envelope();
-        match self.submit_envelope(inner_envelope, scoping, None).await {
+        match self
+            .submit_envelope(inner_envelope, scoping, partition_key)
+            .await
+        {
             Ok(_) => {
                 envelope.accept();
             }
@@ -328,37 +329,27 @@ impl EnvelopeManagerService {
     }
 
     async fn handle_send_metrics(&self, message: SendMetrics) {
-        let SendMetrics {
+        let SendMetrics { buckets, scoping } = message;
+
+        #[allow(unused_mut)]
+        let mut partitions = self.config.metrics_partitions();
+        #[allow(unused_mut)]
+        let mut max_batch_size_bytes = self.config.metrics_max_batch_size_bytes();
+
+        #[cfg(feature = "processing")]
+        if self.store_forwarder.is_some() {
+            // Partitioning on processing relays does not make sense, they end up all
+            // in the same Kafka topic anyways and the partition key is ignored.
+            partitions = None;
+            max_batch_size_bytes = self.config.metrics_max_batch_size_bytes_processing();
+        }
+
+        self.enveloper_processor.send(EncodeMetrics {
             buckets,
             scoping,
-            partition_key,
-        } = message;
-
-        let upstream = self.config.upstream_descriptor();
-        let dsn = PartialDsn {
-            scheme: upstream.scheme(),
-            public_key: scoping.project_key,
-            host: upstream.host().to_owned(),
-            port: upstream.port(),
-            path: "".to_owned(),
-            project_id: Some(scoping.project_id),
-        };
-
-        let mut item = Item::new(ItemType::MetricBuckets);
-        item.set_payload(ContentType::Json, serde_json::to_vec(&buckets).unwrap());
-        let mut envelope = Envelope::from_request(None, RequestMeta::outbound(dsn));
-        envelope.add_item(item);
-
-        let partition_key = partition_key.map(|x| x.to_string());
-        let result = self.submit_envelope(envelope, scoping, partition_key).await;
-        if let Err(err) = result {
-            relay_log::trace!(
-                error = &err as &dyn Error,
-                "failed to submit the envelope, merging buckets back",
-            );
-            self.aggregator
-                .send(MergeBuckets::new(scoping.project_key, buckets));
-        }
+            max_batch_size_bytes,
+            partitions,
+        });
     }
 
     async fn handle_send_client_reports(&self, message: SendClientReports) {